Place partition.hpp in include/utilities/partition/mpi
